Threaded IO: add io-threads-name config and apply pthread_setname on IO threads
With the introduction of "Threaded IO", multiple IO threads may be created.
Redis also creates other threads for non-IO tasks.  By default, all the child
threads inherit the process' name, e.g. "redis-server".

External to the Redis process, it is impossible to identify these child
threads. The main thread is known as that is the parent process.

However, high-performance Redis applications may want to modify the CPU and NUMA
affinity of IO and main threads, but keep the other threads elsewhere.

This change assigns a name to the IO threads using the config `io-threads-name`.
Defaulting to "redis-server-io", we apply it at thread creation time.

Since we must use the non-portable `pthread_setname_np` that is only available
on Linux, and thus we shim it with `redis_pthread_setname`.